+++ CHANGELOG.md
@@ -1,10 +1,29 @@
 ﻿## v1.7.1
 **Fixes:
-- Implemented a hopefully temporary fix to a bug where streamed audio would randomly have the wrong pitch or volume
+- Implemented a hopefully temporary fix for a bug where streamed audio would randomly have the wrong pitch or volume
 - Moved audio streaming to a worker thread and an internal `SoundEffectInstance`-deriving type to fix audio popping at lower framerates
 
 **API Changes:**
 - Removed the internal `FastReflection` class and its related classes in favor of an assembly publicizer package
+- `MonoSoundLibrary`
+  - Obsolete members:
+    - `void Init()`
+      - Replacement: `void Init(Game)`
+      - `MonoSoundLibrary` now has to track the `Game` it's assigned to for some new API features
+- `Controls`
+  - New properties:
+    - `StreamFocusBehavior DefaultStreamFocusBehavior { get; set; }`
+      - Controls the default behavior for `StreamPackage`s when the game is not focused
+      - Defaults to `StreamFocusBehavior.KeepPlaying`
+- `StreamPackage`
+  - New methods:
+    - `void Resume()`
+      - Previously missing method for mirroring of SoundEffectInstance handling
+      - Pauses the stream
+  - New properties:
+    - `StreamFocusBehavior? FocusBehavior { get; set; }`
+      - Acts as an override of `Controls.DefaultStreamFocusBehavior`
+      - Defaults to `null`, which is interpreted as using the value of `Controls.DefaultStreamFocusBehavior`
 
 ## v1.7
  **Fixes:**